replication test with preset round
